requirements.txt
fastapi
uvicorn[standard]
torch
pillow
numpy
python-multipart
ultralytics